Handle disabled Pi Monitor case
