+++ PiStatsMobile/PiStatsWidget/Views/PiMonitorWidgetView.swift
@@ -8,72 +8,6 @@
 import SwiftUI
 import WidgetKit
 
-struct PiMonitorView: View {
-    var provider: PiholeDataProvider
-    let imageSize: CGFloat = 15
-    
-    var body: some View {
-        VStack (alignment:.leading) {
-            VStack (alignment:.leading) {
-                Label(title: {
-                    Text(provider.name)
-                }, icon: {
-                    ViewUtils.shieldStatusImageForDataProvider(provider)
-                    
-                })
-                .font(Font.headline.weight(.bold))
-                Divider()
-                Spacer()
-            }
-            
-            VStack (alignment:.leading, spacing: 6.0) {
-                
-                Label(title: {
-                    Text(provider.temperature)
-                }, icon: {
-                    Image(systemName: UIConstants.SystemImages.metricTemperature)
-                        .frame(width: imageSize, height: imageSize)
-                })
-                .foregroundColor(UIConstants.Colors.domainsOnBlocklist)
-                Label(title: {
-                    Text(provider.uptime)
-                }, icon: {
-                    Image(systemName: UIConstants.SystemImages.metricUptime)
-                        .frame(width: imageSize, height: imageSize)
-                    
-                })
-                .foregroundColor(UIConstants.Colors.queriesBlocked)
-                
-                Label(title: {
-                    Text(provider.memoryUsage)
-                }, icon: {
-                    Image(systemName: UIConstants.SystemImages.metricMemoryUsage)
-                        .frame(width: imageSize, height: imageSize)
-                })
-                .foregroundColor(UIConstants.Colors.totalQueries)
-                                
-                Label(title: {
-                    Text(provider.loadAverage)
-                }, icon: {
-                    Image(systemName: UIConstants.SystemImages.metricLoadAverage)
-                        .frame(width: imageSize, height: imageSize)
-                })
-                .foregroundColor(UIConstants.Colors.percentBlocked)
-            }
-            .font(Font.body.weight(.bold))
-            .minimumScaleFactor(0.89)
-
-            
-        }
-        .frame(
-            maxWidth: .infinity,
-            maxHeight: .infinity,
-            alignment: .topLeading
-        ).padding()
-        .font(.headline)
-    }
-}
-
 struct PiMonitorWidgetView: View {
     var entry: PiholeEntry
     
@@ -82,7 +16,17 @@ struct PiMonitorWidgetView: View {
             UIConstants.Colors.piMonitorWidgetBackground
             if entry.piholeDataProvider.piholes.count == 0 {
                 PiMonitorView(provider: PiholeDataProvider.previewData() ).redacted(reason: .placeholder)
-            } else {
+            } else if entry.piholeDataProvider.canDisplayMetrics == false {
+                VStack (spacing: 10) {
+                    Image(systemName: UIConstants.SystemImages.piholeSetupMonitor)
+                    Text("\(UIConstants.Strings.Widget.piholeNotEnabledOn) \(entry.piholeDataProvider.name)")
+                        .multilineTextAlignment(.center)
+                }
+                .font(Font.headline.weight(.semibold))
+                .foregroundColor(UIConstants.Colors.domainsOnBlocklist)
+                .padding()
+            }
+            else {
                 PiMonitorView(provider: entry.piholeDataProvider)
             }
         }
